move drawer to component
